+++ chapter-4/vk_pipelines.cpp
@@ -34,7 +34,7 @@ VkPipeline PipelineBuilder::build_pipeline(VkDevice device)
     pipelineInfo.sType = VK_STRUCTURE_TYPE_GRAPHICS_PIPELINE_CREATE_INFO;
     pipelineInfo.pNext = &_renderInfo;
 
-    pipelineInfo.stageCount = _shaderStages.size();
+    pipelineInfo.stageCount = (uint32_t)_shaderStages.size();
     pipelineInfo.pStages = _shaderStages.data();
     pipelineInfo.pVertexInputState = &_vertexInputInfo;
     pipelineInfo.pInputAssemblyState = &_inputAssembly;
